Merge with branch 'development'

* Basic spreadsheet functionalities.

+++ src/common/patterns/ClassCounter.tpp
@@ -1,11 +1,31 @@
 // -*- c++ -*- Coloration Syntaxique pour Emacs
+//=====================================================================
+// SimTaDyn: A GIS in a spreadsheet.
+// Copyright 2017 Quentin Quadrat <lecrapouille@gmail.com>
+//
+// This file is part of SimTaDyn.
+//
+// SimTaDyn is free software: you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with GNU Emacs.  If not, see <http://www.gnu.org/licenses/>.
+//=====================================================================
+
 #ifndef CLASS_COUNTER_TPP_
 #  define CLASS_COUNTER_TPP_
 
 #  include <cstddef>
 #  include <atomic>
 
-typedef uint32_t Key;
+typedef size_t Key;
 
 // *************************************************************************************************
 //! \brief This class allows to count the number of created
